only groups creator or admin can add,edit,delete chores

+++ application/chores/views.py
@@ -1,8 +1,9 @@
-from application import app, db
+from application import app, db, login_manager, login_required
 from flask import redirect, render_template, request, url_for
-from flask_login import login_required, current_user
+from flask_login import current_user
 from application.chores.models import Chore
 from application.instances.models import Instance
+from application.groups.models import Group
 from application.chores.forms import ChoreForm
 
 
@@ -14,18 +15,34 @@ def chores_index():
 @app.route("/chores/new/<int:group_id>")
 @login_required
 def chores_form(group_id):
+
+    g = Group.query.get(group_id)
+    if g.creator_id !=  current_user.id and current_user.role != "ADMIN":
+        return login_manager.unauthorized()
+        
     return render_template("chores/new.html", form=ChoreForm(), group_id=group_id)
 
 
 @app.route("/chores/edit/<int:chore_id>/")
 @login_required
 def chores_edit_form(chore_id):
+    c = Chore.query.get(chore_id)
+    g = Group.query.get(c.group_id)
+
+    if g.creator_id !=  current_user.id and current_user.role != "ADMIN":
+        return login_manager.unauthorized()
+
     return render_template("chores/edit.html", chore = Chore.query.get(chore_id), form=ChoreForm())
 
 @app.route("/chores/<int:chore_id>/", methods=["POST"])
 @login_required
 def chores_edit(chore_id):
     c = Chore.query.get(chore_id)
+    g = Group.query.get(c.group_id)
+
+    if g.creator_id !=  current_user.id and current_user.role != "ADMIN":
+        return login_manager.unauthorized()
+
     form = ChoreForm(request.form)
 
     if not form.validate():
@@ -41,6 +58,11 @@ def chores_edit(chore_id):
 def chores_delete(chore_id):
 
     c = Chore.query.get(chore_id)
+    g = Group.query.get(c.group_id)
+
+    if g.creator_id !=  current_user.id and current_user.role != "ADMIN":
+        return login_manager.unauthorized()
+
     Instance.query.filter_by(chore_id = chore_id).delete()
     db.session.delete(c)
     db.session().commit()
@@ -51,6 +73,11 @@ def chores_delete(chore_id):
 @app.route("/chores/group/<int:group_id>", methods=["POST"])
 @login_required
 def chores_create(group_id):
+
+    g = Group.query.get(group_id)
+    if g.creator_id !=  current_user.id and current_user.role != "ADMIN":
+        return login_manager.unauthorized()
+
     form = ChoreForm(request.form)
 
     if not form.validate():
